Move sc test to test_sky_coord
